A cross product of the wrong vectors, and the swizzles around it
A tangent space lighting shader of the ordinary kind, written to see an sm5
pixel shader end to end, found four things - one of them wrong output.

cross(n, t) came out as cross(float3(t3.x, t0.yz), t3.zyx). The cross product
grouper finds its vectors by asking which factor of each product is which, with
a node equality that compared inputs and nothing more: normalize(n).x was
normalize(n).z, and t0.x - a variable, with no inputs at all - was t3.x. Two
components of a multi output node are not one value and two variables are not
one variable; the equality says so now.

Every named value a cross or a normalize read came out permuted, with every
reader swizzled back, because the candidates a value is named from are walked
from the last component back and the variable took them in that order. They are
put in the order the value has them. Three older fixtures lose their
permutations with it.

`flags & 0.000000`: the integer analysis had never been told what a constant
buffer variable is declared as, so an and over a uint had nothing to type its
immediate by. It seeds from the reflection data now. And a register's type was
looked up by register rather than by component, so `uint a; int b;` in one
register made b a uint.

The listing was dropping two things: cb0[r0.x + 1] printed as cb0[0], and an
and over a float's bits printed the float, l(1) for 0x3f800000. Both fixed.

+++ HlslDecompiler.Tests/ShaderSources/ps_4_0_instruction/integer_hash.fx
@@ -6,7 +6,7 @@ float4 main() : SV_Target
 	float4 o;
 
 	int4 r0;
-	r0.x = (int)seed;
+	r0.x = seed;
 	r0.y = 0;
 	while (true) {
 		r0.z = (r0.y >= count) ? -1 : 0;
