+++ HlslDecompiler.Tests/ShaderSources/vs_4_0_instruction/input_array.fx
@@ -4,7 +4,7 @@ float4 main(float4 texcoord[4] : TEXCOORD) : SV_Position
 {
 	float4 o;
 
-	float r0;
+	int r0;
 	r0 = n;
 	o = texcoord[r0.x];
